[move]ShipTypeクラスをlogbook.constantsパッケージへ移動
