+++ src/main/java/logbook/bean/Ship.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 
 import javax.json.JsonObject;
 
+import logbook.constants.ShipType;
 import logbook.internal.JsonHelper;
-import logbook.internal.ShipType;
 import logbook.internal.Ships;
 import logbook.internal.SlotItemType;
 import lombok.Data;
