+++ src/main/java/logbook/internal/Ships.java
@@ -42,6 +42,7 @@ import logbook.bean.SlotitemMstCollection;
 import logbook.bean.Stype;
 import logbook.bean.StypeCollection;
 import logbook.common.Messages;
+import logbook.constants.ShipType;
 import logbook.internal.Tuple.Pair;
 import logbook.plugin.PluginServices;
 import lombok.AllArgsConstructor;
